Long overdue first commit
This project has taken much longer than I thought it would, and I'm also
learning some valuable lessons when  it comes to git. I'm finding that
it is also incredibly important to commit often when working with CSS,
because often times things will break without you realizing, because
you're working on one layout and another one is getting messed up. As of
this commit, my mobile layout is working perfectly, and my tablet and
desktop layouts are not working as intended. The biggest thing that is
killing me is trying to get the text of my header to wrap around the
header image.